[FIX]: Исправлено автоматическое добавление светлой и темной темы при создании таблицы
