+++ packages/server/models/theme.ts
@@ -30,23 +30,4 @@ Theme.init(
   }
 );
 
-export const seedThemes = async () => {
-  try {
-    const lightTheme = await Theme.findOne({ where: { theme: 'light' } });
-    const darkTheme = await Theme.findOne({ where: { theme: 'dark' } });
-
-    if (!lightTheme) {
-      await Theme.create({ theme: 'light', description: 'Light theme' });
-      console.log('Light theme created');
-    }
-
-    if (!darkTheme) {
-      await Theme.create({ theme: 'dark', description: 'Dark theme' });
-      console.log('Dark theme created');
-    }
-  } catch (err) {
-    console.error('Error seeding themes:', err);
-  }
-};
-
 export default Theme;
