+++ packages/server/models/user-theme.ts
@@ -16,9 +16,8 @@ UserTheme.init(
       autoIncrement: true,
       primaryKey: true,
     },
-    theme: {
+    themeName: {
       type: new DataTypes.TEXT(),
-      allowNull: false,
     },
     userId: {
       type: DataTypes.INTEGER.UNSIGNED,
@@ -31,7 +30,7 @@ UserTheme.init(
   }
 );
 
-UserTheme.belongsTo(Theme, { foreignKey: 'theme' });
+UserTheme.belongsTo(Theme, { foreignKey: 'themeName', targetKey: 'theme' });
 UserTheme.belongsTo(User, { foreignKey: 'userId' });
 
 export default UserTheme;
